fixed the bug found in last commit 	modified:   ch10/tree.hpp
	rewrote:

	void recur_add
	(sPointer& current, const ValueType& val, sPointer prev = nullptr)

	added parent operation.
	checking_prev_print(sPointer node) is working now.

+++ ch10/main.cpp
@@ -3,7 +3,7 @@
 
 int main()
 {
-    ch10::tree::binary_tree<long> tree;
+    ch10::tree::binary_tree<int> tree;
 
     for(int i=50; i != 100; ++i)
         tree.insert(i);
